[feat] waiting 삭제
