+++ src/main/java/com/projectw/domain/reservation/controller/ReservationController.java
@@ -1,0 +1,15 @@
+package com.projectw.domain.reservation.controller;
+
+import com.projectw.domain.reservation.service.ReservationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/v1/reservation")
+public class ReservationController {
+
+    private final ReservationService reservationService;
+
+}
