+++ src/main/java/com/projectw/domain/reservation/service/ReservationService.java
@@ -1,0 +1,15 @@
+package com.projectw.domain.reservation.service;
+
+import com.projectw.domain.reservation.repository.ReservationRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Slf4j
+@Service
+public class ReservationService {
+
+    private final ReservationRepository reservationRepository;
+
+}
