--- src/main/java/com/projectw/domain/waiting/repository/WaitingDslRepository.java
@@ -1,4 +1,0 @@
-package com.projectw.domain.waiting.repository;
-
-public interface WaitingDslRepository {
-}
