--- src/main/java/com/projectw/domain/waiting/repository/WaitingRepository.java
@@ -1,9 +1,0 @@
-package com.projectw.domain.waiting.repository;
-
-import com.projectw.domain.waiting.entity.Waiting;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WaitingRepository extends JpaRepository<Waiting, Long>
-        , WaitingDslRepository {
-
-}
